chore: release v0.2.306
Dep bumps:
  - rmcp 1.6 -> 1.7
  - tree-sitter 0.25 -> 0.26.9 (breaking: named_child takes u32 now;
    cast applied in src/utils/bash_parser.rs)
  - assorted patch-level bumps across the lockfile

Full test suite (52 tests) + clippy -D warnings still green.

+++ src/utils/bash_parser.rs
@@ -63,7 +63,7 @@ fn collect_statement_nodes(node: Node<'_>, statements: &mut Vec<StatementNode>) 
         });
     }
 
-    for index in 0..node.named_child_count() {
+    for index in 0..node.named_child_count() as u32 {
         if let Some(child) = node.named_child(index) {
             collect_statement_nodes(child, statements);
         }
